{Style] 소스 코드 정리
코드 보기 편하게 정리했습니다
