--- BattleSystem/BattleSystem.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33424.131
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BattleSystem", "BattleSystem\BattleSystem.csproj", "{E47528C8-9E49-48ED-9123-64CC7CDB4BD5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E47528C8-9E49-48ED-9123-64CC7CDB4BD5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E47528C8-9E49-48ED-9123-64CC7CDB4BD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E47528C8-9E49-48ED-9123-64CC7CDB4BD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E47528C8-9E49-48ED-9123-64CC7CDB4BD5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {5DF62B25-6463-435A-A607-E7087FCC06E0}
-	EndGlobalSection
-EndGlobal
