+++ IPG/IPG.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.13.35919.96 d17.13
+VisualStudioVersion = 17.5.33424.131
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Main", "Main.csproj", "{FD030646-7373-4ED6-8BD4-056137F56AD4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IPG", "IPG\IPG.csproj", "{354C7DAF-2A01-4DB7-9E5A-90F35154AFA8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FD030646-7373-4ED6-8BD4-056137F56AD4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FD030646-7373-4ED6-8BD4-056137F56AD4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FD030646-7373-4ED6-8BD4-056137F56AD4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FD030646-7373-4ED6-8BD4-056137F56AD4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{354C7DAF-2A01-4DB7-9E5A-90F35154AFA8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{354C7DAF-2A01-4DB7-9E5A-90F35154AFA8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{354C7DAF-2A01-4DB7-9E5A-90F35154AFA8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{354C7DAF-2A01-4DB7-9E5A-90F35154AFA8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2A38BB37-1D33-410E-99F1-B8F004B21088}
+		SolutionGuid = {C1CE7685-F31E-4EB6-8C12-4FD49242C8B1}
 	EndGlobalSection
 EndGlobal
