+++ IPG/IPG/Script/MonsterController.cs
@@ -1,19 +1,18 @@
 ﻿using System;
-
 // 네임스페이스 -> 프로그램 클래스 안에 Monster[] monsters 필드 선언, 몬스터 클래스는 프로그램 클래스 밖에 따로 생성
 
-namespace MyApp
+namespace RPG
 {
-    internal class Program
+    internal class ControllMonster
     {
-        static Monster[] monsters = new Monster[]
+        static MonsterController[] monsters = new MonsterController[]
         {
-            new Monster(2, "미니언", 15, 5),
-            new Monster(3, "공허충", 10, 9),
-            new Monster(5, "대포미니언", 25, 8)
+            new MonsterController(2, "미니언", 15, 5),
+            new MonsterController(3, "공허충", 10, 9),
+            new MonsterController(5, "대포미니언", 25, 8)
         };
 
-        static void Main(string[] args)
+        public void Enter()
         {
             while (true)
             {
@@ -51,14 +50,14 @@ namespace MyApp
         }
     }
 
-    internal class Monster
+    internal class MonsterController
     {
         public int Level;
         public string Name;
         public int Hp;
         public int Atk;
 
-        public Monster(int level, string name, int hp, int atk)
+        public MonsterController(int level, string name, int hp, int atk)
         {
             Level = level;
             Name = name;
