+++ IPG/IPG/Script/PlayerController.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RPG
+{
+    internal class PlayerController
+    {
+        public int Level = 1;
+        public string Name = "Chad";
+        public string Job = "전사";
+        public int Atk = 10;
+        public int Def = 5;
+        public int Hp = 100;
+        public int Gold = 1500;
+    }
+}
