+++ IPG/IPG/Script/Start.cs
@@ -1,0 +1,13 @@
+﻿using System.Numerics;
+
+namespace RPG
+{
+    internal class Start
+    {
+        static void Main(string[] args)
+        {
+            VillageController village = new VillageController();
+            village.Enter();
+        }
+    }
+}
