--- Monsters/Monsters.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30907.101
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Monsters", "Monsters.csproj", "{CB574E03-1D37-4765-9B2E-257F37F65CA6}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CB574E03-1D37-4765-9B2E-257F37F65CA6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CB574E03-1D37-4765-9B2E-257F37F65CA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CB574E03-1D37-4765-9B2E-257F37F65CA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CB574E03-1D37-4765-9B2E-257F37F65CA6}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D3C88D1E-3CCF-498E-AC63-5499492EFEFE}
-	EndGlobalSection
-EndGlobal
